move stuff to component
