+++ src/app/[locale]/account/team/[teamId]/(withLeftNav)/integrations/integration/list/[intId]/components/IntegrationHistoryTable/IntegrationHistoryTable.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useState } from "react";
 import {
     Table,
     TableBody,
@@ -9,15 +10,28 @@ import {
     Typography,
 } from "@mui/material";
 import { IntegrationHistory } from "@/interfaces/IntegrationHistory";
+import { PaginationType } from "@/interfaces/Pagination";
+import Box from "@/components/Box";
+import Pagination from "@/components/Pagination";
 import TickCrossIcon from "@/components/TickCrossIcon";
 import { colors } from "@/config/theme";
 
-const IntegrationHistoryTable = ({
-    integrations,
-}: {
-    integrations: Array<IntegrationHistory>;
-}) => {
-    const rows = integrations.map(x => (
+const IntegrationHistoryTable = ({}: {}) => {
+    const [currentPage, setCurrentPage] = useState(1);
+
+    // Example, put data fetch here
+    const data: PaginationType<IntegrationHistory> = {
+        lastPage: 1,
+        to: 1,
+        from: 1,
+        currentPage: 1,
+        total: 1,
+        list: [],
+    };
+
+    const { lastPage, list } = data || {};
+
+    const rows = list.map(x => (
         <TableRow>
             <TableCell>
                 {x.run_time}
@@ -32,11 +46,21 @@ const IntegrationHistoryTable = ({
     ));
 
     return (
-        <TableContainer sx={{ width: "100%" }}>
-            <Table>
-                <TableBody>{rows}</TableBody>
-            </Table>
-        </TableContainer>
+        <Box>
+            <TableContainer sx={{ width: "100%" }}>
+                <Table>
+                    <TableBody>{rows}</TableBody>
+                </Table>
+            </TableContainer>
+            <Pagination
+                isLoading={false}
+                page={currentPage}
+                count={lastPage}
+                onChange={(e: React.ChangeEvent<unknown>, page: number) =>
+                    setCurrentPage(page)
+                }
+            />
+        </Box>
     );
 };
